crypto_arbitrage_bot/requirements.txt
ccxt
pandas
aiohttp
websocket-client
cryptography
pydantic
sqlalchemy
flask
plotly
python-dotenv
uvicorn
fastapi
requests
